Estimate Marginal distribution and Gaussian Copula for Gene Expression Matrix

Estimate Marginal distribution and Gaussian Copula for Gene Expression Matrix

Usage

Est_ModelPara(
  expr,
  anno,
  sim_method = c("ind", "copula"),
  region = NULL,
  ncores = 1,
  subsetN = NULL
)

Arguments

expr

Expression levels of input data

anno

Cell type annotation of input data

sim_method

c('ind', 'copula')

region

description

ncores

No of cores for parallel computing.

subsetN

If the number of cells is large, one can use it to estimate expression distribuitons from subset of cells (e.g. subsetN=2500) per cell type.

Value

Estimated Gaussian Copula