Merge spatial and expression data from multiple regions

Merge spatial and expression data from multiple regions

Usage

MergeRegion(points.list, expr.list)

Arguments

points.list

points.list is a list of points from multiple regions

expr.list

expr.list is a list of expressions from multiple regions

Value

meta:

Output the meta data of all cells in all regions

count:

Output the count data of all cells in all regions