Generate cell location data by modeling the spatial information of existing data.

Poisson random process model separately for each cell type is the foundation. The impact of CCIs and other spatial patterns update the initial cell allocation.

Usage

cell.loc.model.fc(
  n,
  PointLoc,
  PointAnno,
  PointRegion,
  window_method,
  seed = NULL,
  cell.inh.attr.input1 = NULL,
  same.dis.cutoff = 0.01,
  even.distribution.coef = 0
)

Arguments

n

No. of cells

PointLoc

The location of input cells on x, y axis.

PointAnno

The cell type annotation of input cells.

PointRegion

The spatial regions of input cells.

window_method

Method for estimating window of cells. Choices include "network", "convex", "convex2", "convex3", convex5", "rectangle". Default is "network", which is most accurate.

seed

Random seed.

same.dis.cutoff

Cutoff value for regarding two cells are in the same location (default = 0).

even.distribution.coef

A parameter to avoid cells gathering from each other.

Value

Simulated cell spatial map file.