Save cell attraction and inhibition patterns for simulation

Save cell attraction and inhibition patterns for simulation

Usage

cellProximityTable(
  gobject,
  output_file = file.path(getwd(), "est_CCI_dist_dist.csv"),
  abs_enrichm = 0.3,
  p_adj = 0.05,
  spatial_network_name = "Delaunay_network",
  save.unfiltered = F,
  seed = NULL
)

Arguments

gobject

Giotto object, output of preprocessGiotto.

output_file

Provide a path and file name to save the CCI results for use in sCCIgen.

abs_enrichm

Effect size threshold for saving.

p_adj

Adjusted p-value threshold (Default = 0.05) for saving.

spatial_network_name

Networks to choose. Possible values include "Delaunay_network", "distance_based_network", "knn_network".

save.unfiltered

If save a parallel table for all results, unfiltered by asb_enrich and p_adj.

seed

Seed.