Generate model parameters from data.

This function generates model parameters from input data for all cell types in regions.

Usage

generate_marginal_params(expr, feature, region = NULL, ncores = 1)

Arguments

expr

Gene expression level (count).

feature

Cell features (e.g. cell type, spatial coordinates, regions) of reference data.

ncores

No of cores

region_specific_model

Whether estimation model differ in different regions.

Value

Provide model parameters including marginal distributions and copula (if not NULL) for all cell types in all regions.